chore: Rename getElapsedTimeSinceInitialScan -> getElapsedTimeInDaysSinceInitialScan
