+++ src/app/functions/server/getElapsedTimeInDaysSinceInitialScan.ts
